+++ src/main/java/model/entities/impl/EnemyImpl.java
@@ -87,7 +87,7 @@ public final class EnemyImpl implements Enemy {
      * {@inheritDoc}
      */
     @Override
-    public void setX(final int x) {
+    public void setX(final double x) {
         this.x = x;
     }
 
@@ -95,7 +95,7 @@ public final class EnemyImpl implements Enemy {
      * {@inheritDoc}
      */
     @Override
-    public void setY(final int y) {
+    public void setY(final double y) {
         this.y = y;
     }
